refactor(unroll): group analyses under analyses/ (pliron-style)
Move loop_info.rs + induction.rs into mir-transforms/src/analyses/, mirroring
pliron (analyses/liveness.rs, graph/dominance.rs): concept-named files, no
-analysis suffix, the directory marks them as read-only analyses. The transform
(unroll.rs) stays at the crate root, keeping analysis vs transform separated.
