+++ crates/mir-transforms/src/analyses/loop_info.rs
